requirements.txt
imbalanced-learn==0.8.0
tensorflow==2.1.0
matplotlib==3.3.2
seaborn==0.11.0
psutil==5.7.3
tqdm==4.48.0
PyYAML==5.3.1
wandb==0.10.12